Added support to Bro structs for marshaling to/from up to 3-dimensional arrays of primitives. (#226)

+++ rt/robovm/src/main/java/org/robovm/rt/bro/annotation/Array.java
@@ -38,5 +38,5 @@ public @interface Array {
     /**
      * Array dimensions.
      */
-    long[] value();
+    int[] value();
 }
